Removes DOM operations from render functions (moved to main.js)
