+++ src/components/menu.js
@@ -13,16 +13,3 @@ export const createMenuTemplate = (menuNames) => {
     </nav>
   `);
 };
-
-export const setMenuItemActive = (menuElement) => {
-  const menuItems = document.querySelectorAll(`.trip-tabs__btn`);
-
-  for (const menuItem of menuItems) {
-    if (menuItem.classList.contains(`trip-tabs__btn--active`)) {
-      menuItem.classList.remove(`trip-tabs__btn--active`);
-      break;
-    }
-  }
-
-  menuElement.classList.add(`trip-tabs__btn--active`);
-};
